# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jadi.ia API - Plataforma de criação de sites e aplicações com IA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: Listar projetos do usuário
      responses:
        "200":
          description: Lista de projetos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Criar novo projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Projeto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Buscar projeto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Projeto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Atualizar projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Projeto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Deletar projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Projeto deletado

  # FILES
...
  /projects/{id}/secrets:
    get:
      operationId: listProjectSecrets
      tags: [secrets]
      summary: Listar secrets do projeto (apenas chaves)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de secrets (apenas chaves visíveis)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSecret"

    post:
      operationId: createProjectSecret
      tags: [secrets]
      summary: Criar secret no projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretBody"
      responses:
        "201":
          description: Secret criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSecret"
  /projects/{id}/secrets/{secretId}:
    delete:
      operationId: deleteProjectSecret
      tags: [secrets]
      summary: Deletar secret do projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: secretId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Secret deletado

  # AI
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        theme:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - language
        - theme
        - isPublic
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        theme:
          type: string
      required:
        - name
        - language
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        theme:
          type: string
        isPublic:
          type: boolean

    # FILES
...
    ProjectSecret:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        key:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - key
        - createdAt
    CreateSecretBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value

    # AI
...